requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Jinja2==3.1.2
Werkzeug==2.3.6
gunicorn==20.1.0  # If you're deploying with Gunicorn
Pillow==9.1.0  # For image resizing
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
MarkupSafe==2.1.3
requests==2.31.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==2.0.4
WTForms==3.0.1
zipp==3.15.0
